Move test_cuda_apply to directory
